requirements.txt
altair==5.5.0
asttokens==2.1.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
debugpy==1.6.3
decorator==5.1.1
deep-translator==1.11.4
detoxify==0.5.2
entrypoints==0.4
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.3.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==7.4.7
keybert==0.9.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
moviepy==2.1.2
mpmath==1.3.0
narwhals==1.36.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper==20240930
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.11
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyngrok==7.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stack-data==0.6.1
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
whisper==1.1.10
yt-dlp==2025.3.31
